updated beacon format for testbed, and added in the necessary hooks for constructing a payload with all contexts

+++ src/blend_project_templates/Thingy_IoT_SensorKit_v2.1.0/testbed/sensor.h
@@ -24,7 +24,7 @@ typedef struct {
 }humidity_t;
 
 typedef struct {
-  int32_t  integer;
+  int16_t  integer;
   uint8_t  decimal;
   uint32_t timestamp_ms;
 } pressure_t;
